+++ CHANGELOG.md
@@ -12,6 +12,7 @@ For details about compatibility between different releases, see the **Commitment
 ### Added
 
 - Support claim in device import in the Console.
+- New Admin Panel in the Console.
 
 ### Changed